# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "CarRent Support",
            "url": "https://www.carrent.com/support",
            "email": "support@carrent.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/account-role": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get account roles data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account-role"
                ],
                "summary": "Get account roles data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "search by account id",
                        "name": "account_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "search by role id",
                        "name": "role_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort result by attributes",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "must-revalidate",
                            "none"
                        ],
                        "type": "string",
                        "description": "Request Cache Control",
                        "name": "Cache-Control",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AccountRolesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.AccountRolesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.AccountRolesResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Create account role data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account-role"
                ],
                "summary": "Create AccountRole",
                "parameters": [
                    {
                        "description": "AccountRole Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateAccountRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SingleAccountRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.SingleAccountRoleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.SingleAccountRoleResponse"
                        }
                    }
                }
            }
        },
        "/account-role/{id}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get account role by id data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account-role"
                ],
                "summary": "Get account role by id data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "must-revalidate",
                            "none"
                        ],
                        "type": "string",
                        "description": "Request Cache Control",
                        "name": "Cache-Control",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SingleAccountRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.SingleAccountRoleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.SingleAccountRoleResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Delete account role data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account-role"
                ],
                "summary": "Delete account role data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.EmptyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.EmptyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AccountRole": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "role_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "model.AccountRolesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AccountRole"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/model.Pagination"
                },
                "status_code": {
                    "type": "integer"
                },
                "transaction_info": {
                    "$ref": "#/definitions/model.TransactionInfo"
                },
                "translation": {
                    "$ref": "#/definitions/model.Translation"
                }
            }
        },
...
        "model.CreateAccountRole": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.EmptyResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "transaction_info": {
                    "$ref": "#/definitions/model.TransactionInfo"
                },
                "translation": {
                    "$ref": "#/definitions/model.Translation"
                }
            }
        },
...
        "model.Pagination": {
            "type": "object",
            "properties": {
                "current_elements": {
                    "type": "integer"
                },
                "current_page": {
                    "type": "integer"
                },
                "sort_by": {
                    "type": "string"
                },
                "total_elements": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "model.SingleAccountRoleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.AccountRole"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "transaction_info": {
                    "$ref": "#/definitions/model.TransactionInfo"
                },
                "translation": {
                    "$ref": "#/definitions/model.Translation"
                }
            }
        },
...
        "model.TransactionInfo": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "string"
                },
                "request_method": {
                    "type": "string"
                },
                "request_uri": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "model.Translation": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "%s"
        }
    }
}